# 前端/myoj/openapitools.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "通用后端模版 API 文档",
    "description": "登录、注册、用户管理等接口",
    "contact": {
      "name": "开发团队",
      "email": "dev@example.com"
    },
    "version": "1.0"
  },
  "servers": [
    {
      "url": "http://localhost:8888/api",
      "description": "Generated server url"
    }
  ],
...
  "paths": {
...
    "/question/run": {
      "post": {
        "tags": [
          "题目接口"
        ],
        "summary": "运行代码（样例）",
        "description": "调用沙箱执行当前代码与题目公开样例（最多 3 组），不入库、不计榜",
        "operationId": "runQuestionSample",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/QuestionRunRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResultQuestionRunVO"
                }
              }
            }
          }
        }
      }
    },
    "/question/page": {
      "post": {
        "tags": [
          "题目接口"
        ],
        "summary": "公开分页查询题目",
        "description": "题库分页列表（无需管理员）",
        "operationId": "pageQuestions",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/QuestionQueryRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResultPageQuestionAdminListItemVO"
                }
              }
            }
          }
        }
      }
    },
    "/question/admin/update": {
      "post": {
        "tags": [
          "题目接口"
        ],
        "summary": "管理员更新题目",
        "description": "按 id 更新题目",
        "operationId": "adminUpdateQuestion",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/QuestionUpdateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResultString"
                }
              }
            }
          }
        }
      }
    },
    "/question/admin/page": {
      "post": {
        "tags": [
          "题目接口"
        ],
        "summary": "管理员分页查询题目",
        "description": "支持按标题、内容、标签、出题人筛选",
        "operationId": "adminPageQuestions",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/QuestionQueryRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResultPageQuestionAdminListItemVO"
                }
              }
            }
          }
        }
      }
    },
    "/question/admin/delete/batch": {
      "post": {
        "tags": [
          "题目接口"
        ],
        "summary": "管理员批量删除题目",
        "description": "支持单个与批量删除",
        "operationId": "adminBatchDeleteQuestions",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/QuestionBatchDeleteRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResultBoolean"
                }
              }
            }
          }
        }
      }
    },
    "/question/add": {
      "post": {
        "tags": [
          "题目接口"
        ],
        "summary": "创建题目",
        "description": "创建题目",
        "operationId": "addQuestion",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/QuestionAddRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResultLong"
                }
              }
            }
          }
        }
      }
    },
...
    "/question/public/get/{id}": {
      "get": {
        "tags": [
          "题目接口"
        ],
        "summary": "公开获取题目详情",
        "description": "用于题库进入作答页",
        "operationId": "getPublicQuestionDetail",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResultQuestionPublicDetailVO"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "QuestionRunRequest": {
        "type": "object",
        "properties": {
          "questionId": {
            "type": "string",
            "minLength": 1,
            "pattern": "\\d{1,24}"
          },
          "language": {
            "type": "string",
            "minLength": 1
          },
          "code": {
            "type": "string",
            "maxLength": 200000,
            "minLength": 0
          }
        },
        "required": [
          "code",
          "language",
          "questionId"
        ]
      },
      "JudgeInfo": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string"
          },
          "memory": {
            "type": "integer",
            "format": "int64"
          },
          "time": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
      "QuestionRunCaseVO": {
        "type": "object",
        "properties": {
          "index": {
            "type": "integer",
            "format": "int32"
          },
          "input": {
            "type": "string"
          },
          "expectedOutput": {
            "type": "string"
          },
          "actualOutput": {
            "type": "string"
          },
          "passed": {
            "type": "boolean"
          }
        }
      },
      "QuestionRunVO": {
        "type": "object",
        "properties": {
          "sandboxCode": {
            "type": "integer",
            "format": "int32"
          },
          "sandboxMessage": {
            "type": "string"
          },
          "judgeInfo": {
            "$ref": "#/components/schemas/JudgeInfo"
          },
          "terminalError": {
            "type": "boolean"
          },
          "cases": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/QuestionRunCaseVO"
            }
          },
          "allSamplePassed": {
            "type": "boolean"
          },
          "warning": {
            "type": "string"
          },
          "hints": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "suspectedShortcut": {
            "type": "boolean"
          }
        }
      },
      "ResultQuestionRunVO": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "format": "int32"
          },
          "message": {
            "type": "string"
          },
          "data": {
            "$ref": "#/components/schemas/QuestionRunVO"
          },
          "timestamp": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
      "QuestionQueryRequest": {
        "type": "object",
        "properties": {
          "current": {
            "type": "integer",
            "format": "int64"
          },
          "pageSize": {
            "type": "integer",
            "format": "int64"
          },
          "sortField": {
            "type": "string"
          },
          "sortOrder": {
            "type": "string"
          },
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "title": {
            "type": "string"
          },
          "tags": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "content": {
            "type": "string"
          },
          "userId": {
            "type": "integer",
            "format": "int64"
          },
          "userNickname": {
            "type": "string"
          }
        }
      },
      "JudgeCase": {
        "type": "object",
        "properties": {
          "input": {
            "type": "string",
            "maxLength": 20000,
            "minLength": 0
          },
          "output": {
            "type": "string",
            "maxLength": 20000,
            "minLength": 0
          }
        },
        "required": [
          "input",
          "output"
        ]
      },
      "JudgeConfig": {
        "type": "object",
        "properties": {
          "timeLimit": {
            "type": "integer",
            "format": "int64",
            "maximum": 10000
          },
          "memoryLimit": {
            "type": "integer",
            "format": "int64",
            "maximum": 1048576
          },
          "stackLimit": {
            "type": "integer",
            "format": "int64",
            "maximum": 262144
          }
        },
        "required": [
          "memoryLimit",
          "stackLimit",
          "timeLimit"
        ]
      },
      "OrderItem": {
        "type": "object",
        "properties": {
          "column": {
            "type": "string"
          },
          "asc": {
            "type": "boolean"
          }
        }
      },
      "PageQuestionAdminListItemVO": {
        "type": "object",
        "properties": {
          "records": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/QuestionAdminListItemVO"
            }
          },
          "total": {
            "type": "integer",
            "format": "int64"
          },
          "size": {
            "type": "integer",
            "format": "int64"
          },
          "current": {
            "type": "integer",
            "format": "int64"
          },
          "orders": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/OrderItem"
            },
            "writeOnly": true
          },
          "optimizeCountSql": {
            "writeOnly": true
          },
          "searchCount": {
            "writeOnly": true
          },
          "optimizeJoinOfCountSql": {
            "type": "boolean",
            "writeOnly": true
          },
          "maxLimit": {
            "type": "integer",
            "format": "int64",
            "writeOnly": true
          },
          "countId": {
            "type": "string",
            "writeOnly": true
          },
          "pages": {
            "type": "integer",
            "format": "int64",
            "deprecated": true
          }
        }
      },
      "QuestionAdminListItemVO": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "title": {
            "type": "string"
          },
          "content": {
            "type": "string"
          },
          "tags": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "answer": {
            "type": "string"
          },
          "submitNum": {
            "type": "integer",
            "format": "int32"
          },
          "acceptedNum": {
            "type": "integer",
            "format": "int32"
          },
          "judgeCase": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/JudgeCase"
            }
          },
          "judgeConfig": {
            "$ref": "#/components/schemas/JudgeConfig"
          },
          "thumbNum": {
            "type": "integer",
            "format": "int32"
          },
          "favourNum": {
            "type": "integer",
            "format": "int32"
          },
          "userId": {
            "type": "integer",
            "format": "int64"
          },
          "userNickname": {
            "type": "string"
          },
          "createTime": {
            "type": "string",
            "format": "date-time"
          },
          "updateTime": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "ResultPageQuestionAdminListItemVO": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "format": "int32"
          },
          "message": {
            "type": "string"
          },
          "data": {
            "$ref": "#/components/schemas/PageQuestionAdminListItemVO"
          },
          "timestamp": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
      "QuestionUpdateRequest": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "title": {
            "type": "string",
            "maxLength": 255,
            "minLength": 0
          },
          "content": {
            "type": "string",
            "maxLength": 20000,
            "minLength": 0
          },
          "tags": {
            "type": "array",
            "items": {
              "type": "string",
              "minLength": 1
            },
            "minItems": 1
          },
          "answer": {
            "type": "string",
            "maxLength": 20000,
            "minLength": 0
          },
          "judgeCase": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/JudgeCase"
            },
            "minItems": 1
          },
          "judgeConfig": {
            "$ref": "#/components/schemas/JudgeConfig"
          }
        },
        "required": [
          "answer",
          "content",
          "id",
          "judgeCase",
          "judgeConfig",
          "tags",
          "title"
        ]
      },
      "ResultString": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "format": "int32"
          },
          "message": {
            "type": "string"
          },
          "data": {
            "type": "string"
          },
          "timestamp": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
      "QuestionBatchDeleteRequest": {
        "type": "object",
        "properties": {
          "ids": {
            "type": "array",
            "items": {
              "type": "integer",
              "format": "int64"
            },
            "minItems": 1
          }
        },
        "required": [
          "ids"
        ]
      },
      "ResultBoolean": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "format": "int32"
          },
          "message": {
            "type": "string"
          },
          "data": {
            "type": "boolean"
          },
          "timestamp": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
      "QuestionAddRequest": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "maxLength": 255,
            "minLength": 0
          },
          "content": {
            "type": "string",
            "minLength": 1
          },
          "tags": {
            "type": "array",
            "items": {
              "type": "string",
              "minLength": 1
            },
            "maxItems": 10,
            "minItems": 0
          },
          "answer": {
            "type": "string",
            "minLength": 1
          },
          "judgeCase": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/JudgeCase"
            },
            "maxItems": 20,
            "minItems": 0
          },
          "judgeConfig": {
            "$ref": "#/components/schemas/JudgeConfig"
          }
        },
        "required": [
          "answer",
          "content",
          "judgeCase",
          "judgeConfig",
          "tags",
          "title"
        ]
      },
      "ResultLong": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "format": "int32"
          },
          "message": {
            "type": "string"
          },
          "data": {
            "type": "integer",
            "format": "int64"
          },
          "timestamp": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
...
      "QuestionPublicDetailVO": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "title": {
            "type": "string"
          },
          "content": {
            "type": "string"
          },
          "tags": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "submitNum": {
            "type": "integer",
            "format": "int32"
          },
          "acceptedNum": {
            "type": "integer",
            "format": "int32"
          },
          "judgeConfig": {
            "$ref": "#/components/schemas/JudgeConfig"
          },
          "sampleJudgeCase": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/JudgeCase"
            }
          },
          "userId": {
            "type": "integer",
            "format": "int64"
          },
          "userNickname": {
            "type": "string"
          },
          "createTime": {
            "type": "string",
            "format": "date-time"
          },
          "updateTime": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "ResultQuestionPublicDetailVO": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "format": "int32"
          },
          "message": {
            "type": "string"
          },
          "data": {
            "$ref": "#/components/schemas/QuestionPublicDetailVO"
          },
          "timestamp": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
...
    }
  }
}